pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnivoice-server"
version = "0.1.1"
description = "OpenAI-compatible HTTP server for OmniVoice TTS"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "zamery", email = "matthew.ngo1114@gmail.com" }]

keywords = ["tts", "text-to-speech", "omnivoice", "openai", "fastapi", "voice-cloning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # FIX: pin omnivoice to a compatible minor series.
    # OmniVoice is brand-new (April 2026) and its API may change in 0.2.x.
    # Widen this bound only after reviewing upstream changelogs.
    "omnivoice>=0.1.0,<0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "platformdirs>=3.0.0",
    "psutil>=6.0.0",
    "huggingface-hub>=0.26.0",
    "soundfile>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/maemreyo/omnivoice-server"
Documentation = "https://github.com/maemreyo/omnivoice-server#readme"
Repository = "https://github.com/maemreyo/omnivoice-server"
Issues = "https://github.com/maemreyo/omnivoice-server/issues"
Changelog = "https://github.com/maemreyo/omnivoice-server/blob/main/CHANGELOG.md"

[project.optional-dependencies]
benchmark = ["psutil", "tqdm"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "torchcodec>=0.11",
]

[project.scripts]
omnivoice-server = "omnivoice_server.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["omnivoice_server"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["omnivoice_server"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
precision = 1
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
